2022/LogRhythm Case Management Search APP/requirements.txt
alembic==1.7.7
altgraph==0.17.2
bcrypt==4.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.3.post0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.19.5
openpyxl==3.0.9
pandas==1.1.5
pefile==2021.9.3
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.42
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0